changed datatype from float to double

+++ dqops/src/integration-test/java/com/dqops/spark/sensors/column/strings/SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSpecIntegrationTest.java
@@ -64,7 +64,7 @@ public class SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSp
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 86.666, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 86.666, (double) resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -77,7 +77,7 @@ public class SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSp
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 86.666, (float)  resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 86.666, (double)  resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -90,7 +90,7 @@ public class SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSp
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 86.666, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 86.666, (double) resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -103,7 +103,7 @@ public class SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSp
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(25, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(83.333, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals(83.333, (double) resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -116,6 +116,6 @@ public class SparkColumnStringsStringBooleanPlaceholderPercentSensorParametersSp
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 86.666, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 86.666, (double) resultTable.column(0).get(0), 0.001);
     }
 }
