+++ dqops/src/integration-test/java/com/dqops/spark/sensors/column/strings/SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegrationTest.java
@@ -64,7 +64,7 @@ public class SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegra
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(65.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(65.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -77,7 +77,7 @@ public class SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegra
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(65.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(65.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -90,7 +90,7 @@ public class SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegra
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(65.0f, resultTable.column(0).get(0));
+        Assertions.assertEquals(65.0d, resultTable.column(0).get(0));
     }
 
     @Test
@@ -103,7 +103,7 @@ public class SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegra
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(6, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 0.0, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 0.0, (double) resultTable.column(0).get(0), 0.001);
     }
 
     @Test
@@ -116,6 +116,6 @@ public class SparkColumnStringsStringValidUuidPercentSensorParametersSpecIntegra
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(6, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals((float) 0.0, (float) resultTable.column(0).get(0), 0.001);
+        Assertions.assertEquals((double) 0.0, (double) resultTable.column(0).get(0), 0.001);
     }
 }
